requirements.txt
fastapi
uvicorn
psycopg2-binary
asyncio
pydantic
pandas